Migrating from TruncatableView (custom code) to ViewThatFits (Apple-provided solution)
