+++ phoenix-ios/phoenix-ios/Localizable.xcstrings
@@ -20921,7 +20921,7 @@
         "de" : {
           "stringUnit" : {
             "state" : "translated",
-            "value" : "Wenn du [Advanced Data Protection] (https://support.apple.com/en-us/108756) für iCloud aktivierst, wird deine Wiederherstellungsphrase Ende-zu-Ende verschlüsselt und Apple hat keinen Zugriff darauf."
+            "value" : "Wenn du [Advanced Data Protection](https://support.apple.com/en-us/108756) für iCloud aktivierst, wird deine Wiederherstellungsphrase Ende-zu-Ende verschlüsselt und Apple hat keinen Zugriff darauf."
           }
         },
         "es" : {
