+++ phoenix-ios/phoenix-ios/views/configuration/privacy and security/recovery phrase/RecoveryPhraseView.swift
@@ -778,257 +778,3 @@ fileprivate struct SyncErrorDetails: View, ViewName {
 		waiting.skip()
 	}
 }
-
-fileprivate struct RecoveryPhraseReveal: View {
-	
-	@Binding var isShowing: Bool
-	
-	let recoveryPhrase: RecoveryPhrase
-	let language: MnemonicLanguage
-	
-	@State var truncationDetected: Bool = false
-	
-	@StateObject var toast = Toast()
-	
-	@Environment(\.colorScheme) var colorScheme: ColorScheme
-	
-	init(isShowing: Binding<Bool>, recoveryPhrase: RecoveryPhrase) {
-		self._isShowing = isShowing
-		self.recoveryPhrase = recoveryPhrase
-		self.language = recoveryPhrase.language ?? MnemonicLanguage.english
-	}
-	
-	func mnemonic(_ idx: Int) -> String {
-		let mnemonics = recoveryPhrase.mnemonicsArray
-		return (mnemonics.count > idx) ? mnemonics[idx] : " "
-	}
-	
-	@ViewBuilder
-	var body: some View {
-		
-		ZStack {
-			GeometryReader { geometry in
-				ScrollView(.vertical) {
-					content()
-						.frame(width: geometry.size.width)
-						.frame(minHeight: geometry.size.height)
-				}
-			}
-			
-			header()
-			toast.view()
-		}
-	}
-	
-	@ViewBuilder
-	func header() -> some View {
-		
-		VStack {
-			HStack(alignment: VerticalAlignment.center, spacing: 0) {
-				Text(verbatim: "\(language.flag) \(language.displayName)")
-					.font(.callout)
-					.foregroundColor(.secondary)
-				Spacer()
-				Button {
-					close()
-				} label: {
-					Image("ic_cross")
-						.resizable()
-						.frame(width: 30, height: 30)
-				}
-			}
-			Spacer()
-		}
-		.padding()
-	}
-	
-	@ViewBuilder
-	func content() -> some View {
-		
-		VStack {
-			
-			Spacer()
-			
-			Text("KEEP THIS SEED SAFE.")
-				.font(.title2)
-				.lineLimit(nil)
-				.multilineTextAlignment(.center)
-				.fixedSize(horizontal: false, vertical: true)
-				.padding(.bottom, 2)
-			Text("DO NOT SHARE.")
-				.font(.title2)
-				.lineLimit(nil)
-				.multilineTextAlignment(.center)
-				.fixedSize(horizontal: false, vertical: true)
-			
-			Spacer()
-			
-			Group {
-				if truncationDetected {
-					singleColumnLayout()
-				} else {
-					twoColumnLayout()
-				}
-			}
-			.environment(\.layoutDirection, .leftToRight) // issue #237
-			.padding(.top, 20)
-			.padding(.bottom, 10)
-			
-			Spacer()
-			Spacer()
-			
-			copyButton()
-				.padding(.bottom, 6)
-
-			Text("BIP39 seed with standard BIP84 derivation path")
-				.font(.footnote)
-				.foregroundColor(.secondary)
-				.fixedSize(horizontal: false, vertical: true)
-			
-		}
-		.padding(.top, 20)
-		.padding([.leading, .trailing], 30)
-		.padding(.bottom, 20)
-	}
-	
-	@ViewBuilder
-	func twoColumnLayout() -> some View {
-		
-		HStack(alignment: VerticalAlignment.center, spacing: 0) {
-			Spacer()
-			
-			VStack {
-				ForEach(0..<6, id: \.self) { idx in
-					Text(verbatim: "#\(idx + 1) ")
-						.font(.headline)
-						.lineLimit(1)
-						.foregroundColor(.secondary)
-						.padding(.bottom, 2)
-				}
-			}
-			.padding(.trailing, 2)
-			
-			VStack(alignment: .leading) {
-				ForEach(0..<6, id: \.self) { idx in
-					TruncatableView(
-						fixedHorizontal: true,
-						fixedVertical: true
-					) {
-						Text(mnemonic(idx))
-							.font(.headline)
-							.lineLimit(1)
-					} wasTruncated: {
-						self.wasTruncated(visibleIdx: idx+1)
-					}
-					.padding(.bottom, 2)
-				} // </ForEach>
-			} // </VStack>
-			.padding(.trailing, 4) // boost spacing a wee bit
-			
-			Spacer()
-			
-			VStack {
-				ForEach(6..<12, id: \.self) { idx in
-					Text(verbatim: "#\(idx + 1) ")
-						.font(.headline)
-						.lineLimit(1)
-						.foregroundColor(.secondary)
-						.padding(.bottom, 2)
-				}
-			}
-			.padding(.trailing, 2)
-			
-			VStack(alignment: .leading) {
-				ForEach(6..<12, id: \.self) { idx in
-					TruncatableView(
-						fixedHorizontal: true,
-						fixedVertical: true
-					) {
-						Text(mnemonic(idx))
-							.font(.headline)
-							.lineLimit(1)
-					} wasTruncated: {
-						self.wasTruncated(visibleIdx: idx+1)
-					}
-					.padding(.bottom, 2)
-				} // </ForEach>
-			} // </VStack>
-			
-			Spacer()
-		}
-	}
-	
-	@ViewBuilder
-	func singleColumnLayout() -> some View {
-		
-		HStack(alignment: VerticalAlignment.center, spacing: 0) {
-			
-			VStack(alignment: HorizontalAlignment.leading, spacing: 0) {
-				ForEach(0..<12, id: \.self) { idx in
-					Text(verbatim: "#\(idx + 1) ")
-						.font(.headline)
-						.foregroundColor(.secondary)
-						.padding(.bottom, 2)
-				} // </ForEach>
-			} // </VStack>
-			.padding(.trailing, 2)
-			
-			VStack(alignment: HorizontalAlignment.leading, spacing: 0) {
-				ForEach(0..<12, id: \.self) { idx in
-					Text(mnemonic(idx))
-						.font(.headline)
-						.padding(.bottom, 2)
-				} // </ForEach>
-			} // </VStack>
-			
-		} // </HStack>
-	}
-	
-	@ViewBuilder
-	func copyButton() -> some View {
-		
-		HStack(alignment: VerticalAlignment.center, spacing: 0) {
-			Spacer()
-				
-			Button {
-				copyRecoveryPhrase()
-			} label: {
-				Text("Copy").font(.title3)
-			}
-			
-			Spacer()
-		} // </HStack>
-	}
-
-	
-	func wasTruncated(visibleIdx: Int) {
-		log.trace("[RecoverySeedReveal] wasTruncated(#: \(visibleIdx))")
-		
-		DispatchQueue.main.async {
-			truncationDetected = true
-		}
-	}
-	
-	func copyRecoveryPhrase() {
-		log.trace("[RecoverySeedReveal] copyRecoveryPhrase()")
-		
-		copy(recoveryPhrase.mnemonics)
-	}
-	
-	private func copy(_ string: String) {
-		log.trace("[RecoverySeedReveal] copy()")
-		
-		UIPasteboard.general.string = string
-		AppDelegate.get().clearPasteboardOnReturnToApp = true
-		toast.pop(
-			"Pasteboard will be cleared when you return to Phoenix.",
-			colorScheme: colorScheme.opposite,
-			duration: 4.0 // seconds
-		)
-	}
-	
-	func close() {
-		log.trace("[RecoverySeedReveal] close()")
-		isShowing = false
-	}
-}
