+++ phoenix-ios/phoenix-ios/views/inspect/Details/Details_Outgoing_Lightning.swift
@@ -21,6 +21,8 @@ struct Details_Outgoing_Lightning: DetailsInfoGrid {
 	
 	@ObservedObject var currencyPrefs = CurrencyPrefs.current
 	
+	@Environment(\.dynamicTypeSize) var dynamicTypeSize: DynamicTypeSize
+	
 	// --------------------------------------------------
 	// MARK: View Builders
 	// --------------------------------------------------
