+++ phoenix-ios/phoenix-ios/views/inspect/Details/Details_Outgoing_SpliceCpfp.swift
@@ -21,6 +21,8 @@ struct Details_Outgoing_SpliceCpfp: DetailsInfoGrid {
 	
 	@ObservedObject var currencyPrefs = CurrencyPrefs.current
 	
+	@Environment(\.dynamicTypeSize) var dynamicTypeSize: DynamicTypeSize
+	
 	// --------------------------------------------------
 	// MARK: View Builders
 	// --------------------------------------------------
